+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -13,6 +13,7 @@ group: Deployment & Configs
 ### RC 
 
 - Fixed 'Add Rule-based Filter' name is getting an unexpected integer added in creation. [#754](https://github.com/mcneel/rhino.inside-revit/issues/754)
+- Added 'Rule-based Filter Definiton' component.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}